Revert "feat: email plugins support (#2611)"
This reverts commit d8b55d30f97bec7376d6fc4d8b449d7137864782.
