+++ .vscode/settings.json
@@ -6,12 +6,5 @@
     "packages/workflow-core",
     "services/workflows-service",
     "packages/common"
-  ],
-  "search.exclude": {
-    "**/node_modules": true,
-    "**/dist": true,
-    "**/data-migrations": false
-  },
-  "search.followSymlinks": true,
-  "search.useIgnoreFiles": false
+  ]
 }
