+++ apps/kyb-app/package.json
@@ -5,7 +5,6 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "start": "vite",
     "build": "tsc && vite build",
     "lint": "eslint . --fix",
     "format": "prettier --write .",
