+++ packages/workflow-core/src/lib/plugins/common-plugin/iterative-plugin.test.ts
@@ -80,7 +80,7 @@ describe('workflow-runner', () => {
         server.close();
       });
 
-      it.skip('transitions to successAction and persist response to context', async () => {
+      it('transitions to successAction and persist response to context', async () => {
         const workflow = createWorkflowRunner(
           definition,
           // @ts-expect-error - see the comments on `webhookPluginsSchemas`
