+++ packages/workflow-core/src/lib/plugins/external-plugin/kyb-plugin.ts
@@ -20,9 +20,7 @@ export class KybPlugin extends ApiPlugin {
       companyNumber = '',
       vendor = 'open-corporates',
     } = payload;
-
-    let countryCode: string | undefined;
-
+    let countryCode: string;
     if (
       typeof countryOfIncorporation === 'string' &&
       countryOfIncorporation.length === 2 &&
@@ -35,7 +33,6 @@ export class KybPlugin extends ApiPlugin {
 
     if (typeof countryCode !== 'string')
       throw new Error('Invalid countryOfIncorporation for KYB process');
-
     if (typeof companyNumber !== 'string') throw new Error('Invalid companyNumber for KYB process');
 
     const jurisdictionCode =
