+++ packages/workflow-core/src/lib/plugins/external-plugin/kyc-session-plugin.ts
@@ -2,7 +2,6 @@ import { AnyRecord } from '@ballerine/common';
 import { ApiPlugin } from './api-plugin';
 import { JsonSchemaValidator } from '../../utils/context-validator/json-schema-validator';
 import { Validator } from '../../utils';
-import { BallerineApiPlugin } from './ballerine-plugin';
 
 const kycSessionRequestSchema = {
   $schema: 'http://json-schema.org/draft-07/schema#',
@@ -24,8 +23,9 @@ const kycSessionRequestSchema = {
   required: ['firstName', 'lastName', 'callbackUrl', 'vendor'],
 };
 
-export class KycSessionPlugin extends BallerineApiPlugin {
+export class KycSessionPlugin extends ApiPlugin {
   public static pluginType = 'http';
+  public static pluginKind = 'kyc-session';
 
   async validateContent<TValidationContext extends 'Request' | 'Response'>(
     schemaValidator: Validator | undefined,
@@ -49,11 +49,10 @@ export class KycSessionPlugin extends BallerineApiPlugin {
     payload: AnyRecord,
     headers: HeadersInit,
   ) {
-    const callbackUrlWithPlaceholder = await this.replaceAllVariables(
+    const callbackUrlWithPlaceholder = await this.replaceValuePlaceholders(
       payload['callbackUrl'] as string,
       payload,
     );
-
     const callbackUrl = new URL(callbackUrlWithPlaceholder);
     callbackUrl.searchParams.set('processName', 'kyc-unified-api');
     payload['callbackUrl'] = callbackUrl.toString();
