+++ packages/workflow-core/src/lib/plugins/external-plugin/types.ts
@@ -4,7 +4,6 @@ import { ActionablePlugin } from '../types';
 
 import { AnyRecord } from '@ballerine/common';
 import { SecretsManager } from '@/lib/types';
-import { ApiEmailTemplates } from './vendor-consts';
 
 export interface ValidatableTransformer {
   transformers?: Transformers;
@@ -14,9 +13,8 @@ export interface IApiPluginParams {
   name: string;
   pluginKind?: string;
   stateNames: string[];
-  url: string | { url: string; options: Record<string, string> };
+  url: string;
   vendor?: string;
-  template?: ApiEmailTemplates;
   method: 'POST' | 'PUT' | 'PATCH' | 'DELETE' | 'GET';
   request?: ValidatableTransformer;
   response?: ValidatableTransformer;
