+++ packages/workflow-core/src/lib/utils/definition-validator/extensions-validator.ts
@@ -1,19 +1,16 @@
 import type { StateMachine } from 'xstate';
 import type {
   IDispatchEventPluginParams,
+  ISerializableChildPluginParams,
   ISerializableHttpPluginParams,
   SerializableIterativePluginParams,
   SerializableValidatableTransformer,
   SerializableWebhookPluginParams,
 } from '../../plugins/external-plugin/types';
-import { BALLERINE_API_PLUGINS_KINDS } from './../../plugins/external-plugin/vendor-consts';
-
-import type { ISerializableChildPluginParams } from '../../plugins/common-plugin/types';
-
-import type { DispatchEventPlugin } from '@/lib/plugins';
-import { isErrorWithMessage } from '@ballerine/common';
 import { WorkflowEvents, WorkflowExtensions } from '../../types';
 import { ruleValidator } from './rule-validator';
+import { isErrorWithMessage } from '@ballerine/common';
+import type { DispatchEventPlugin } from '@/lib/plugins';
 
 export const extensionsValidator = (
   extensions: WorkflowExtensions,
@@ -25,11 +22,9 @@ export const extensionsValidator = (
     if (
       pluginKind === 'api' ||
       pluginKind === 'kyb' ||
+      pluginKind === 'kyc-session' ||
       pluginKind === 'email' ||
-      pluginKind === 'kyc' ||
-      BALLERINE_API_PLUGINS_KINDS.includes(
-        pluginKind as (typeof BALLERINE_API_PLUGINS_KINDS)[number],
-      )
+      pluginKind === 'kyc'
     ) {
       validateApiPlugin(plugin as unknown as ISerializableHttpPluginParams, states);
     }
