+++ services/workflows-service/scripts/workflows/dynamic-ui-workflow.ts
@@ -42,7 +42,7 @@ export const dynamicUiWorkflowDefinition = {
       collection_invite: {
         on: {
           INVITATION_SENT: 'collection_flow',
-          INVITATION_FAILURE: 'failed',
+          INVIATION_FAILURE: 'failed',
         },
       },
       collection_flow: {
@@ -145,47 +145,165 @@ export const dynamicUiWorkflowDefinition = {
   extensions: {
     apiPlugins: [
       {
-        name: 'invitation-email',
-        pluginKind: 'template-email',
-        template: 'invitation',
+        name: 'collection_invite_email',
+        pluginKind: 'email',
+        url: `{secret.EMAIL_API_URL}`,
         successAction: 'INVITATION_SENT',
-        errorAction: 'INVITATION_FAILURE',
-        stateNames: ['idle', 'collection_invite'],
+        errorAction: 'INVIATION_FAILURE',
+        method: 'POST',
+        stateNames: ['collection_invite'],
+        headers: {
+          Authorization: 'Bearer {secret.EMAIL_API_TOKEN}',
+          'Content-Type': 'application/json',
+        },
+        request: {
+          transform: [
+            {
+              transformer: 'jmespath',
+              mapping: `{
+                customerName: metadata.customerName,
+                collectionFlowUrl: join('',['{secret.COLLECTION_FLOW_URL}','/?token=',metadata.token,'&lng=',workflowRuntimeConfig.language]),
+                from: 'no-reply@ballerine.com',
+                receivers: [entity.data.additionalInfo.mainRepresentative.email],
+                language: workflowRuntimeConfig.language,
+                templateId: 'd-8949519316074e03909042cfc5eb4f02',
+                adapter: '{secret.MAIL_ADAPTER}'
+              }`, // jmespath
+            },
+          ],
+        },
+        response: {
+          transform: [],
+        },
       },
       {
-        name: 'businessInformation',
-        vendor: 'asia-verify',
-        pluginKind: 'registry-information',
-        displayName: 'Registry Information',
+        name: 'kyb',
+        pluginKind: 'api',
+        url: `{secret.UNIFIED_API_URL}/companies-v2/{entity.data.country}/{entity.data.registrationNumber}`,
+        method: 'GET',
         stateNames: ['run_vendor_data'],
         successAction: 'VENDOR_DONE',
         errorAction: 'VENDOR_FAILED',
+        persistResponseDestination: 'pluginsOutput.businessInformation',
+        headers: { Authorization: 'Bearer {secret.UNIFIED_API_TOKEN}' },
+        request: {
+          transform: [
+            {
+              transformer: 'jmespath',
+              mapping: `merge(
+                { vendor: 'asia-verify' },
+                entity.data.country == 'HK' && {
+                  callbackUrl: join('',['{secret.APP_API_URL}/api/v1/external/workflows/',workflowRuntimeId,'/hook/VENDOR_DONE','?resultDestination=pluginsOutput.businessInformation.data&processName=kyb-unified-api'])
+                }
+              )`, // jmespath
+            },
+          ],
+        },
+        response: {
+          transform: [
+            {
+              transformer: 'jmespath',
+              mapping: '@', // jmespath
+            },
+          ],
+        },
       },
       {
-        name: 'companySanctions',
-        vendor: 'asia-verify',
-        pluginKind: 'company-sanctions',
+        name: 'company_sanctions',
+        pluginKind: 'api',
+        url: `{secret.UNIFIED_API_URL}/companies/{entity.data.country}/{entity.data.companyName}/sanctions`,
+        method: 'GET',
         stateNames: ['run_vendor_data'],
-        displayName: 'Company Sanctions',
-        errorAction: 'VENDOR_DONE',
-        successAction: 'VENDOR_FAILED',
+        successAction: 'VENDOR_DONE',
+        errorAction: 'VENDOR_FAILED',
+        persistResponseDestination: 'pluginsOutput.companySanctions',
+        headers: { Authorization: 'Bearer {secret.UNIFIED_API_TOKEN}' },
+        request: {
+          transform: [
+            {
+              transformer: 'jmespath',
+              mapping: `{
+                vendor: 'asia-verify'
+              }`, // jmespath
+            },
+          ],
+        },
+        response: {
+          transform: [
+            {
+              transformer: 'jmespath',
+              mapping: '@', // jmespath
+            },
+          ],
+        },
       },
       {
         name: 'ubo',
-        vendor: 'asia-verify',
-        pluginKind: 'ubo',
+        pluginKind: 'api',
+        url: `{secret.UNIFIED_API_URL}/companies/{entity.data.country}/{entity.data.registrationNumber}/ubo`,
+        method: 'GET',
         stateNames: ['run_vendor_data'],
-        displayName: 'UBO Check',
-        errorAction: 'VENDOR_DONE',
         successAction: 'VENDOR_DONE',
+        errorAction: 'VENDOR_FAILED',
+        persistResponseDestination: 'pluginsOutput.ubo.request',
+        headers: { Authorization: 'Bearer {secret.UNIFIED_API_TOKEN}' },
+        request: {
+          transform: [
+            {
+              transformer: 'jmespath',
+              mapping: `{
+                vendor: 'asia-verify',
+                callbackUrl: join('',['{secret.APP_API_URL}/api/v1/external/workflows/',workflowRuntimeId,'/hook/VENDOR_DONE','?resultDestination=pluginsOutput.ubo.data&processName=ubo-unified-api'])
+              }`, // jmespath
+            },
+          ],
+        },
+        response: {
+          transform: [
+            {
+              transformer: 'jmespath',
+              mapping: '{request: @}', // jmespath
+            },
+          ],
+        },
       },
       {
-        name: 'resubmission-email',
-        pluginKind: 'template-email',
-        template: 'resubmission',
+        name: 'resubmission_email',
+        pluginKind: 'email',
+        url: `{secret.EMAIL_API_URL}`,
+        method: 'POST',
         successAction: 'EMAIL_SENT',
         errorAction: 'EMAIL_FAILURE',
         stateNames: ['pending_resubmission'],
+        headers: {
+          Authorization: 'Bearer {secret.EMAIL_API_TOKEN}',
+          'Content-Type': 'application/json',
+        },
+        request: {
+          transform: [
+            {
+              transformer: 'jmespath',
+              // #TODO: create new token (new using old one)
+              mapping: `{
+                kybCompanyName: entity.data.companyName,
+                customerCompanyName: metadata.customerName,
+                firstName: entity.data.additionalInfo.mainRepresentative.firstName,
+                resubmissionLink: join('',['{secret.COLLECTION_FLOW_URL}','/?token=',metadata.token,'&lng=',workflowRuntimeConfig.language]),
+                supportEmail: join('',['support@',metadata.customerName,'.com']),
+                from: 'no-reply@ballerine.com',
+                name: join(' ',[metadata.customerName,'Team']),
+                receivers: [entity.data.additionalInfo.mainRepresentative.email],
+                templateId: 'd-7305991b3e5840f9a14feec767ea7301',
+                revisionReason: documents[].decision[].revisionReason | [0],
+                language: workflowRuntimeConfig.language,
+                adapter: '${env.MAIL_ADAPTER}'
+              }`, // jmespath
+            },
+          ],
+        },
+        response: {
+          transform: [],
+        },
       },
     ],
     childWorkflowPlugins: [
