+++ services/workflows-service/scripts/workflows/kyb-kyc-workflow-definition.ts
@@ -128,7 +128,7 @@ export const kybKycWorkflowDefinition = {
             {
               transformer: 'jmespath',
               mapping: `{
-              templateId: 'd-8949519316074e03909042cfc5eb4f02',
+              templateId: 'd-00a0d5d14cb14fbb9034b53c6ef7e5fa',
               adapter: '${env.MAIL_ADAPTER}'
               from: 'no-reply@ballerine.com',
               receivers: [mainRepresentative.email],
@@ -175,12 +175,38 @@ export const kybKycWorkflowDefinition = {
         },
       },
       {
-        name: 'resubmission-email',
-        pluginKind: 'template-email',
-        template: 'resubmission',
-        successAction: 'EMAIL_SENT',
-        errorAction: 'EMAIL_FAILURE',
+        name: 'resubmission_email',
+        pluginKind: 'email',
+        url: `{secret.EMAIL_API_URL}`,
+        method: 'POST',
         stateNames: ['pending_resubmission'],
+        headers: {
+          Authorization: 'Bearer {secret.EMAIL_API_TOKEN}',
+          'Content-Type': 'application/json',
+        },
+        request: {
+          transform: [
+            {
+              transformer: 'jmespath',
+              mapping: `{
+              kybCompanyName: entity.data.companyName,
+              customerCompanyName: entity.data.additionalInfo.ubos[0].entity.data.additionalInfo.customerCompany,
+              firstName: entity.data.additionalInfo.mainRepresentative.firstName,
+              resubmissionLink: join('',['https://',entity.data.additionalInfo.ubos[0].entity.data.additionalInfo.normalizedCustomerCompany,'.demo.ballerine.app','/workflowRuntimeId=',workflowRuntimeId,'?resubmitEvent=RESUBMITTED']),
+              supportEmail: join('',[entity.data.additionalInfo.ubos[0].entity.data.additionalInfo.normalizedCustomerCompany,'@support.com']),
+              from: 'no-reply@ballerine.com',
+              name: join(' ',[entity.data.additionalInfo.ubos[0].entity.data.additionalInfo.customerCompany,'Team']),
+              receivers: [entity.data.additionalInfo.mainRepresentative.email],
+              templateId: 'd-7305991b3e5840f9a14feec767ea7301',
+              revisionReason: documents[].decision[].revisionReason | [0],
+              adapter: '${env.MAIL_ADAPTER}'
+              }`, // jmespath
+            },
+          ],
+        },
+        response: {
+          transform: [],
+        },
       },
     ],
     childWorkflowPlugins: [
