+++ services/workflows-service/src/auth/session-serializer.ts
@@ -46,7 +46,6 @@ export class SessionSerializer extends PassportSerializer {
           firstName: true,
           lastName: true,
           avatarUrl: true,
-          lastActiveAt: true,
           userToProjects: { select: { projectId: true } },
         },
       });
