+++ services/workflows-service/src/rule-engine/core/test/data-helper.ts
@@ -57,7 +57,7 @@ export const context = {
           lastAnnualGeneralMeetingDate: '2024-02-27',
         },
       ],
-      name: 'businessInformation',
+      name: 'kyb',
       status: 'SUCCESS',
       orderId: 'av202406121547221341867814',
       invokedAt: 1718178448068,
@@ -114,7 +114,7 @@ export const context = {
           matchedFields: ['PrimaryName'],
         },
       ],
-      name: 'companySanctions',
+      name: 'company_sanctions',
       status: 'SUCCESS',
       invokedAt: 1716447914675,
     },
